src/sentiment/sentiment_analyzer.py
"""
Sentiment analysis module for cryptocurrency market data.
"""
from typing import Dict, List, Optional, Tuple
import pandas as pd
import numpy as np
from datetime import datetime, timedelta

class SentimentAnalyzer:
    """Analyzes sentiment data to generate insights and signals."""
    
    def __init__(self, config: Optional[Dict] = None):
        """
        Initialize the sentiment analyzer.
        
        Args:
            config (Dict, optional): Configuration parameters for sentiment analysis
        """
        self.config = config or {
            'sentiment_threshold': 0.5,  # Threshold for significant sentiment
            'volume_weight': 0.3,        # Weight given to trading volume
            'trend_window': 24,          # Hours to consider for trend
            'signal_threshold': 0.7,     # Threshold for generating signals
            'significant_change': 0.1    # Threshold for significant change
        }
    
    def calculate_sentiment_score(self, data: pd.DataFrame) -> float:
        """
        Calculate an overall sentiment score from raw sentiment data.
        
        Args:
            data (pd.DataFrame): DataFrame containing sentiment data
            
        Returns:
            float: Calculated sentiment score
            
        Raises:
            ValueError: If data is empty or missing required columns
        """
        # Validate input data
        if data.empty:
            raise ValueError("Input DataFrame is empty")
            
        required_columns = {'sentiment', 'volume'}
        if not required_columns.issubset(data.columns):
            raise ValueError(f"Input DataFrame missing required columns: {required_columns}")
        
        # Calculate basic sentiment score (average of sentiment values)
        return data['sentiment'].mean()
    
    def analyze_sentiment_trend(self, data: pd.DataFrame) -> Dict:
        """
        Analyze sentiment trends over time.
        
        Args:
            data (pd.DataFrame): DataFrame containing sentiment data with timestamp
            
        Returns:
            Dict: Trend analysis results containing:
                - trend: 'upward', 'downward', or 'neutral'
                - is_significant: boolean indicating if change is significant
                - change_rate: rate of change over the period
                
        Raises:
            ValueError: If data is empty or missing required columns
        """
        # Validate input data
        if data.empty:
            raise ValueError("Input DataFrame is empty")
            
        required_columns = {'timestamp', 'sentiment', 'volume'}
        if not required_columns.issubset(data.columns):
            raise ValueError(f"Input DataFrame missing required columns: {required_columns}")
        
        # Sort data by timestamp
        data = data.sort_values('timestamp')
        
        # Get the window size from config
        window_size = self.config.get('trend_window', 12)
        significant_change = self.config.get('significant_change', 0.1)
        
        # Calculate the trend over the specified window
        if len(data) >= window_size:
            recent_data = data.tail(window_size)
        else:
            recent_data = data
        
        # Calculate linear regression to determine trend
        x = np.arange(len(recent_data))
        y = recent_data['sentiment'].values
        slope, _ = np.polyfit(x, y, 1)
        
        # Calculate total change rate
        start_sentiment = recent_data['sentiment'].iloc[0]
        end_sentiment = recent_data['sentiment'].iloc[-1]
        change_rate = (end_sentiment - start_sentiment) / abs(start_sentiment) if start_sentiment != 0 else end_sentiment
        
        # Determine trend direction and significance
        if abs(change_rate) < significant_change:
            trend = 'neutral'
            is_significant = False
            change_rate = 0  # For neutral trends, set change rate to 0
        else:
            trend = 'upward' if slope > 0 else 'downward'
            is_significant = True
        
        return {
            'trend': trend,
            'is_significant': is_significant,
            'change_rate': change_rate
        }
    
    def calculate_volume_weighted_sentiment(self, data: pd.DataFrame) -> pd.Series:
        """
        Calculate sentiment scores weighted by trading volume.
        
        Args:
            data (pd.DataFrame): DataFrame containing sentiment and volume data
            
        Returns:
            pd.Series: Volume-weighted sentiment scores
        """
        raise NotImplementedError
    
    def generate_sentiment_signals(self, data: pd.DataFrame) -> pd.DataFrame:
        """
        Generate trading signals based on sentiment analysis.
        
        Args:
            data (pd.DataFrame): DataFrame containing sentiment and price data
            
        Returns:
            pd.DataFrame: Generated trading signals
        """
        raise NotImplementedError
    
    def detect_sentiment_thresholds(self, data: pd.DataFrame) -> List[Dict]:
        """
        Detect when sentiment crosses significant thresholds.
        
        Args:
            data (pd.DataFrame): DataFrame containing sentiment data
            
        Returns:
            List[Dict]: List of threshold crossing events
        """
        raise NotImplementedError
