tests/sentiment/test_sentiment_analyzer.py
import unittest
from unittest.mock import Mock, patch
import pandas as pd
from datetime import datetime, timedelta
import numpy as np

class TestSentimentAnalyzer(unittest.TestCase):
    def setUp(self):
        """Set up test fixtures before each test method."""
        self.mock_data = {
            'time': [datetime.now() - timedelta(hours=i) for i in range(24)],
            'sentiment': [0.5 + (i * 0.02) for i in range(24)],  # Simulated sentiment scores
            'volume': [1000 + (i * 100) for i in range(24)]      # Simulated trading volume
        }
        self.test_df = pd.DataFrame(self.mock_data)
    
    def test_calculate_sentiment_score(self):
        """Test that sentiment scores are calculated correctly."""
        from src.sentiment.sentiment_analyzer import SentimentAnalyzer
        
        analyzer = SentimentAnalyzer()
        
        # Test case 1: Basic sentiment calculation
        test_data = pd.DataFrame({
            'sentiment': [0.8, 0.6, 0.4, 0.7],  # Raw sentiment scores
            'volume': [1000, 2000, 1500, 3000]  # Trading volumes
        })
        
        expected_score = 0.625  # Average of sentiment scores
        actual_score = analyzer.calculate_sentiment_score(test_data)
        self.assertAlmostEqual(expected_score, actual_score, places=3)
        
        # Test case 2: Empty DataFrame
        empty_data = pd.DataFrame(columns=['sentiment', 'volume'])
        with self.assertRaises(ValueError):
            analyzer.calculate_sentiment_score(empty_data)
        
        # Test case 3: Missing required columns
        invalid_data = pd.DataFrame({'price': [100, 200]})
        with self.assertRaises(ValueError):
            analyzer.calculate_sentiment_score(invalid_data)
    
    def test_sentiment_trend_analysis(self):
        """Test that sentiment trends are analyzed correctly."""
        from src.sentiment.sentiment_analyzer import SentimentAnalyzer
        
        analyzer = SentimentAnalyzer({
            'trend_window': 12,  # 12-hour window for trend analysis
            'significant_change': 0.1  # 10% change is considered significant
        })
        
        # Test case 1: Upward trend
        upward_trend_data = pd.DataFrame({
            'timestamp': pd.date_range(end=datetime.now(), periods=24, freq='h'),
            'sentiment': [0.3 + i * 0.02 for i in range(24)],  # Gradually increasing sentiment
            'volume': [1000 + i * 100 for i in range(24)]
        })
        
        trend_result = analyzer.analyze_sentiment_trend(upward_trend_data)
        self.assertEqual(trend_result['trend'], 'upward')
        self.assertTrue(trend_result['is_significant'])
        self.assertGreater(trend_result['change_rate'], 0)
        
        # Test case 2: Downward trend
        downward_trend_data = pd.DataFrame({
            'timestamp': pd.date_range(end=datetime.now(), periods=24, freq='h'),
            'sentiment': [0.8 - i * 0.02 for i in range(24)],  # Gradually decreasing sentiment
            'volume': [1000 + i * 100 for i in range(24)]
        })
        
        trend_result = analyzer.analyze_sentiment_trend(downward_trend_data)
        self.assertEqual(trend_result['trend'], 'downward')
        self.assertTrue(trend_result['is_significant'])
        self.assertLess(trend_result['change_rate'], 0)
        
        # Test case 3: Sideways/Neutral trend
        neutral_trend_data = pd.DataFrame({
            'timestamp': pd.date_range(end=datetime.now(), periods=24, freq='h'),
            'sentiment': [0.5 + (np.sin(i/4) * 0.05) for i in range(24)],  # Oscillating sentiment
            'volume': [1000 + i * 100 for i in range(24)]
        })
        
        trend_result = analyzer.analyze_sentiment_trend(neutral_trend_data)
        self.assertEqual(trend_result['trend'], 'neutral')
        self.assertFalse(trend_result['is_significant'])
        self.assertAlmostEqual(trend_result['change_rate'], 0, places=1)
        
        # Test case 4: Empty data
        empty_data = pd.DataFrame(columns=['timestamp', 'sentiment', 'volume'])
        with self.assertRaises(ValueError):
            analyzer.analyze_sentiment_trend(empty_data)
        
        # Test case 5: Missing required columns
        invalid_data = pd.DataFrame({'price': [100, 200]})
        with self.assertRaises(ValueError):
            analyzer.analyze_sentiment_trend(invalid_data)
    
    def test_volume_weighted_sentiment(self):
        """Test that volume-weighted sentiment is calculated correctly."""
        # TODO: Implement test
        pass
    
    def test_sentiment_signals(self):
        """Test that trading signals are generated correctly from sentiment."""
        # TODO: Implement test
        pass
    
    def test_sentiment_threshold_detection(self):
        """Test detection of significant sentiment changes."""
        # TODO: Implement test
        pass

if __name__ == '__main__':
    unittest.main()
